faster isapa code

+++ crypto_aead/isapa128av20/opt_32_armv67m/crypto_aead.c
@@ -4,15 +4,15 @@
 #include "isap.h"
 
 int crypto_aead_encrypt(
-	unsigned char *c,unsigned long *clen,
-	const unsigned char *m,unsigned long mlen,
-	const unsigned char *ad,unsigned long adlen,
+	unsigned char *c, unsigned long long *clen,
+	const unsigned char *m, unsigned long long mlen,
+	const unsigned char *ad, unsigned long long adlen,
 	const unsigned char *nsec,
 	const unsigned char *npub,
 	const unsigned char *k
 ){
-    (void)nsec;
-
+    (void)nsec;  
+    
 	// Ciphertext length is mlen + tag length
 	*clen = mlen+ISAP_TAG_SZ;
 
@@ -26,10 +26,10 @@ int crypto_aead_encrypt(
 }
 
 int crypto_aead_decrypt(
-	unsigned char *m,unsigned long *mlen,
+	unsigned char *m, unsigned long long *mlen,
 	unsigned char *nsec,
-	const unsigned char *c,unsigned long clen,
-	const unsigned char *ad,unsigned long adlen,
+	const unsigned char *c, unsigned long long clen,
+	const unsigned char *ad, unsigned long long adlen,
 	const unsigned char *npub,
 	const unsigned char *k
 ){
